sync

+++ src/wte/wte_main.hpp
@@ -58,7 +58,7 @@ class wte_main final {
             display = al_create_display(engine_cfg::get<int>("screen_width"), engine_cfg::get<int>("screen_height"));
             if(!display) throw std::runtime_error("Failed to configure display!");
 
-            main_timer = al_create_timer(1.0 / engine_cfg::get<float>("TICKS_PER_SECOND"));
+            main_timer = al_create_timer(1.0 / engine_cfg::get<double>("TICKS_PER_SECOND"));
             if(!main_timer) throw std::runtime_error("Failed to create timer!");
 
             main_queue = al_create_event_queue();
